+++ R/normalized.diff.idx.r
@@ -7,10 +7,8 @@
 #' difference among then is in the wavebands used.
 #'
 #' @param spct an R object
-#' @param plus.w.band waveband objects The waveband determine the
-#'   region of the spectrum used in the calculations
-#' @param minus.w.band waveband objects The waveband determine the
-#'   region of the spectrum used in the calculations
+#' @param plus.w.band,minus.w.band waveband objects The wavebands determine the
+#'   regions of the spectrum used in the calculations.
 #' @param f function used for integration taking spct as first argument and a
 #'   list of wavebands as second argument.
 #' @param ... additional arguments passed to f
@@ -28,12 +26,15 @@
 #'   can be used if there is a good reason for it. In every case \code{spct}
 #'   should be of the class expected by \code{f}. When using two wavebands of
 #'   different widths do consider passing to \code{f} a suitable \code{quantity}
-#'   argument. Wavebands can describe weighting functions if desired.
+#'   argument, for example to compare averages rather than integrals. Wavebands
+#'   can describe weighting functions if desired.
+#'
+#'   \deqn{\mathrm{NDxI} = \frac{f(s, wb_\mathrm{plus}) - f(s, wb_\mathrm{minus})}{f(s, wb_\mathrm{plus}) + f(s, wb_\mathrm{minus})}}
 #'
 #' @note Some NDxI indexes are directly based on satellite instrument data, such
 #'   as those in the Landsat satellites. To simulate such indexes using spectral
-#'   reflectande as input, \code{waveband} definitions provided by package
-#'   'photobiologyWavebands' can be used.
+#'   reflectande as input, constructors of \code{waveband} definitions from package
+#'   'photobiologyWavebands' can be useful.
 #'
 #' @export
 #'