+++ NEWS.md
@@ -8,6 +8,8 @@ editor_options:
 
 - Fix bug in `subset2mspct()` when applied to the case of a collection
   containing a single spct object containing a single spectrum.
+- Add specialization of `smooth_spct()` method for class `"cps_spct"`.
+- Add `e_fraction()` and `q_fraction`.
 
 # photobiology 0.10.16